pyproject.toml
[tool.poetry]
name = "spectraverse"
version = "0.1.0"
description = "A modular framework for identifying molecules from spectra and generating spectra from molecules."
authors = ["David Baugher"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2"
pytorch-lightning = "^2.2"
wandb = "^0.16"
hydra-core = "^1.3"
numpy = "*"
pandas = "*"
scikit-learn = "*"

[tool.poetry.dev-dependencies]
pytest = "^8.2"
black = "^24.4"
flake8 = "^7.0"
isort = "^5.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"